pyproject.toml
[project]
name = "obsidian-note-taking-assistant"
version = "0.1.0"
description = "Second Brain RAG for any Obsidian Vault. Works locally with DuckDB"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.3",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.0",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]
